requirements.txt
openai-agents
python-dotenv
pydantic
fastapi
uvicorn   